[Decomp][Spell] Cohesion split: extract target-list assembly to SpellTargetList.cpp
