+++ src/game/WorldHandlers/Spell.cpp
@@ -635,720 +635,6 @@ Spell::~Spell()
 }
 
 /**
- * @brief Builds the spell target lists for each active effect.
- */
-void Spell::FillTargetMap()
-{
-    // TODO: ADD the correct target FILLS!!!!!!
-
-    UnitList tmpUnitLists[MAX_EFFECT_INDEX];                // Stores the temporary Target Lists for each effect
-    uint8 effToIndex[MAX_EFFECT_INDEX] = {0, 1, 2};         // Helper array, to link to another tmpUnitList, if the targets for both effects match
-    for (int i = 0; i < MAX_EFFECT_INDEX; ++i)
-    {
-        SpellEffectEntry const* spellEffect = m_spellInfo->GetSpellEffect(SpellEffectIndex(i));
-        if (!spellEffect)
-        {
-            continue;
-        }
-
-        // not call for empty effect.
-        // Also some spells use not used effect targets for store targets for dummy effect in triggered spells
-        if (spellEffect->Effect == SPELL_EFFECT_NONE)
-        {
-            continue;
-        }
-
-        // targets for TARGET_SCRIPT_COORDINATES (A) and TARGET_SCRIPT
-        // for TARGET_FOCUS_OR_SCRIPTED_GAMEOBJECT (A) all is checked in Spell::CheckCast and in Spell::CheckItem
-        // filled in Spell::CheckCast call
-        if (spellEffect->EffectImplicitTargetA == TARGET_SCRIPT_COORDINATES ||
-            spellEffect->EffectImplicitTargetA == TARGET_FOCUS_OR_SCRIPTED_GAMEOBJECT ||
-            (spellEffect->EffectImplicitTargetA == TARGET_SCRIPT && spellEffect->EffectImplicitTargetB != TARGET_SELF) ||
-            (spellEffect->EffectImplicitTargetB == TARGET_SCRIPT && spellEffect->EffectImplicitTargetA != TARGET_SELF))
-            continue;
-
-        // TODO: find a way so this is not needed?
-        // for area auras always add caster as target (needed for totems for example)
-        if (IsAreaAuraEffect(spellEffect->Effect))
-        {
-            AddUnitTarget(m_caster, SpellEffectIndex(i));
-        }
-
-        // no double fill for same targets
-        for (int j = 0; j < i; ++j)
-        {
-            SpellEffectEntry const* spellEffect1 = m_spellInfo->GetSpellEffect(SpellEffectIndex(j));
-            if (!spellEffect1)
-            {
-                continue;
-            }
-
-            // Check if same target, but handle i.e. AreaAuras different
-            if (spellEffect->EffectImplicitTargetA == spellEffect1->EffectImplicitTargetA && spellEffect->EffectImplicitTargetB == spellEffect1->EffectImplicitTargetB
-                && spellEffect1->Effect != SPELL_EFFECT_NONE
-                && !IsAreaAuraEffect(spellEffect->Effect) && !IsAreaAuraEffect(spellEffect1->Effect))
-                // Add further conditions here if required
-            {
-                effToIndex[i] = j;                          // effect i has same targeting list as effect j
-                break;
-            }
-        }
-
-        if (effToIndex[i] == i)                             // New target combination
-        {
-            // TargetA/TargetB dependent from each other, we not switch to full support this dependences
-            // but need it support in some know cases
-            switch(spellEffect->EffectImplicitTargetA)
-            {
-                case TARGET_NONE:
-                    switch(spellEffect->EffectImplicitTargetB)
-                    {
-                        case TARGET_NONE:
-                            if (m_caster->GetObjectGuid().IsPet())
-                            {
-                                SetTargetMap(SpellEffectIndex(i), TARGET_SELF, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            }
-                            else
-                            {
-                                SetTargetMap(SpellEffectIndex(i), TARGET_EFFECT_SELECT, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            }
-                            break;
-                        default:
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetB, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                    }
-                    break;
-                case TARGET_SELF:
-                    switch(spellEffect->EffectImplicitTargetB)
-                    {
-                        case TARGET_NONE:                   // Fill Target based on A only
-                        case TARGET_EFFECT_SELECT:
-                        case TARGET_SCRIPT:                 // B-target only used with CheckCast here
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetA, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                        case TARGET_AREAEFFECT_INSTANT:         // use B case that not dependent from from A in fact
-                            if ((m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION) == 0)
-                            {
-                                m_targets.setDestination(m_caster->GetPositionX(), m_caster->GetPositionY(), m_caster->GetPositionZ());
-                            }
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetB, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                        case TARGET_BEHIND_VICTIM:              // use B case that not dependent from from A in fact
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetB, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                        default:
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetA, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetB, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                    }
-                    break;
-                case TARGET_EFFECT_SELECT:
-                    switch(spellEffect->EffectImplicitTargetB)
-                    {
-                        case TARGET_NONE:
-                        case TARGET_EFFECT_SELECT:
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetA, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                        // dest point setup required
-                        case TARGET_AREAEFFECT_INSTANT:
-                        case TARGET_AREAEFFECT_CUSTOM:
-                        case TARGET_ALL_ENEMY_IN_AREA:
-                        case TARGET_ALL_ENEMY_IN_AREA_INSTANT:
-                        case TARGET_ALL_ENEMY_IN_AREA_CHANNELED:
-                        case TARGET_ALL_FRIENDLY_UNITS_IN_AREA:
-                        case TARGET_AREAEFFECT_GO_AROUND_DEST:
-                        case TARGET_RANDOM_NEARBY_DEST:
-                            // triggered spells get dest point from default target set, ignore it
-                            if (!(m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION) || m_IsTriggeredSpell)
-                                if (WorldObject* castObject = GetCastingObject())
-                                {
-                                    m_targets.setDestination(castObject->GetPositionX(), castObject->GetPositionY(), castObject->GetPositionZ());
-                                }
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetB, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                        // target pre-selection required
-                        case TARGET_INNKEEPER_COORDINATES:
-                        case TARGET_TABLE_X_Y_Z_COORDINATES:
-                        case TARGET_CASTER_COORDINATES:
-                        case TARGET_SCRIPT_COORDINATES:
-                        case TARGET_CURRENT_ENEMY_COORDINATES:
-                        case TARGET_DUELVSPLAYER_COORDINATES:
-                        case TARGET_DYNAMIC_OBJECT_COORDINATES:
-                        case TARGET_POINT_AT_NORTH:
-                        case TARGET_POINT_AT_SOUTH:
-                        case TARGET_POINT_AT_EAST:
-                        case TARGET_POINT_AT_WEST:
-                        case TARGET_POINT_AT_NE:
-                        case TARGET_POINT_AT_NW:
-                        case TARGET_POINT_AT_SE:
-                        case TARGET_POINT_AT_SW:
-                            // need some target for processing
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetA, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetB, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                        default:
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetB, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                    }
-                    break;
-                case TARGET_CASTER_COORDINATES:
-                    switch(spellEffect->EffectImplicitTargetB)
-                    {
-                        case TARGET_ALL_ENEMY_IN_AREA:
-                            // Note: this hack with search required until GO casting not implemented
-                            // environment damage spells already have around enemies targeting but this not help in case nonexistent GO casting support
-                            // currently each enemy selected explicitly and self cast damage
-                            if (spellEffect->Effect == SPELL_EFFECT_ENVIRONMENTAL_DAMAGE)
-                            {
-                                if (m_targets.getUnitTarget())
-                                {
-                                    tmpUnitLists[i /*==effToIndex[i]*/].push_back(m_targets.getUnitTarget());
-                                }
-                            }
-                            else
-                            {
-                                SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetA, tmpUnitLists[i /*==effToIndex[i]*/]);
-                                SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetB, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            }
-                            break;
-                        case 0:
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetA, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            tmpUnitLists[i /*==effToIndex[i]*/].push_back(m_caster);
-                            break;
-                        default:
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetA, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetB, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                    }
-                    break;
-                case TARGET_TABLE_X_Y_Z_COORDINATES:
-                    switch(spellEffect->EffectImplicitTargetB)
-                    {
-                        case 0:
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetA, tmpUnitLists[i /*==effToIndex[i]*/]);
-
-
-                        // need some target for processing
-                            SetTargetMap(SpellEffectIndex(i), TARGET_EFFECT_SELECT, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                        case TARGET_AREAEFFECT_INSTANT:         // All 17/7 pairs used for dest teleportation, A processed in effect code
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetB, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                        default:
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetA, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetB, tmpUnitLists[i /*==effToIndex[i]*/]);
-                    }
-                    case TARGET_SELF2:
-                    switch(spellEffect->EffectImplicitTargetB)
-                    {
-                        case TARGET_NONE:
-                        case TARGET_EFFECT_SELECT:
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetA, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                        case TARGET_AREAEFFECT_CUSTOM:
-                            // triggered spells get dest point from default target set, ignore it
-                            if (!(m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION) || m_IsTriggeredSpell)
-                                if (WorldObject* castObject = GetCastingObject())
-                                {
-                                    m_targets.setDestination(castObject->GetPositionX(), castObject->GetPositionY(), castObject->GetPositionZ());
-                                }
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetB, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                        // most A/B target pairs is self->negative and not expect adding caster to target list
-                        default:
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetB, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                    }
-                    break;
-                case TARGET_DUELVSPLAYER_COORDINATES:
-                    switch(spellEffect->EffectImplicitTargetB)
-                    {
-                        case TARGET_NONE:
-                        case TARGET_EFFECT_SELECT:
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetA, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            if (Unit* currentTarget = m_targets.getUnitTarget())
-                            {
-                                tmpUnitLists[i /*==effToIndex[i]*/].push_back(currentTarget);
-                            }
-                            break;
-                        default:
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetA, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetB, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                    }
-                    break;
-                case TARGET_SCRIPT:
-                    switch (spellEffect->EffectImplicitTargetB)
-                    {
-                        case TARGET_SELF:
-                            // Fill target based on B only, A is only used with CheckCast here.
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetB, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                        default:
-                            break;
-                    }
-                    break;
-                default:
-                    switch(spellEffect->EffectImplicitTargetB)
-                    {
-                        case 0:
-                        case TARGET_EFFECT_SELECT:
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetA, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                        case TARGET_SCRIPT_COORDINATES:         // B case filled in CheckCast but we need fill unit list base at A case
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetA, tmpUnitLists[i /*==effToIndex[i]*/]);
-                           break;
-                        default:
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetA, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            SetTargetMap(SpellEffectIndex(i), spellEffect->EffectImplicitTargetB, tmpUnitLists[i /*==effToIndex[i]*/]);
-                            break;
-                    }
-                    break;
-            }
-        }
-
-        if (m_caster->GetTypeId() == TYPEID_PLAYER)
-        {
-            Player* me = (Player*)m_caster;
-            for (UnitList::const_iterator itr = tmpUnitLists[effToIndex[i]].begin(); itr != tmpUnitLists[effToIndex[i]].end(); ++itr)
-            {
-                Player* targetOwner = (*itr)->GetCharmerOrOwnerPlayerOrPlayerItself();
-                if (targetOwner && targetOwner != me && targetOwner->IsPvP() && !me->IsInDuelWith(targetOwner))
-                {
-                    me->UpdatePvP(true);
-                    me->RemoveAurasWithInterruptFlags(AURA_INTERRUPT_FLAG_ENTER_PVP_COMBAT);
-                    break;
-                }
-            }
-        }
-
-        for (UnitList::iterator itr = tmpUnitLists[effToIndex[i]].begin(); itr != tmpUnitLists[effToIndex[i]].end();)
-        {
-            if (!CheckTarget(*itr, SpellEffectIndex(i)))
-            {
-                itr = tmpUnitLists[effToIndex[i]].erase(itr);
-                continue;
-            }
-            else
-            {
-                ++itr;
-            }
-        }
-
-        for (UnitList::const_iterator iunit = tmpUnitLists[effToIndex[i]].begin(); iunit != tmpUnitLists[effToIndex[i]].end(); ++iunit)
-        {
-            AddUnitTarget((*iunit), SpellEffectIndex(i));
-        }
-    }
-}
-
-/**
- * @brief Prepares proc-trigger metadata for the current spell cast.
- */
-void Spell::prepareDataForTriggerSystem()
-{
-    //==========================================================================================
-    // Now fill data for trigger system, need know:
-    // an spell trigger another or not ( m_canTrigger )
-    // Create base triggers flags for Attacker and Victim ( m_procAttacker and  m_procVictim)
-    //==========================================================================================
-    // Fill flag can spell trigger or not
-    // TODO: possible exist spell attribute for this
-    m_canTrigger = false;
-
-    if (m_CastItem)
-    {
-        m_canTrigger = false;                               // Do not trigger from item cast spell
-    }
-    else if (!m_IsTriggeredSpell)
-    {
-        m_canTrigger = true;                                // Normal cast - can trigger
-    }
-    else if (!m_triggeredByAuraSpell)
-    {
-        m_canTrigger = true;                                // Triggered from SPELL_EFFECT_TRIGGER_SPELL - can trigger
-    }
-
-    if (!m_canTrigger)                                      // Exceptions (some periodic triggers)
-    {
-        switch (m_spellInfo->GetSpellFamilyName())
-        {
-            case SPELLFAMILY_MAGE:
-                // Arcane Missles / Blizzard triggers need do it
-                if (m_spellInfo->IsFitToFamilyMask(UI64LIT(0x0000000000200080)))
-                {
-                    m_canTrigger = true;
-                }
-                // Clearcasting trigger need do it
-                else if (m_spellInfo->IsFitToFamilyMask(UI64LIT(0x0000000200000000), 0x00000008))
-                {
-                    m_canTrigger = true;
-                }
-                // Replenish Mana, item spell with triggered cases (Mana Agate, etc mana gems)
-                else if (m_spellInfo->IsFitToFamilyMask(UI64LIT(0x0000010000000000)))
-                {
-                    m_canTrigger = true;
-                }
-                break;
-            case SPELLFAMILY_WARLOCK:
-                // For Hellfire Effect / Rain of Fire / Seed of Corruption triggers need do it
-                if (m_spellInfo->IsFitToFamilyMask(UI64LIT(0x0000800000000060)))
-                {
-                    m_canTrigger = true;
-                }
-                break;
-            case SPELLFAMILY_PRIEST:
-                // For Penance,Mind Sear,Mind Flay heal/damage triggers need do it
-                if (m_spellInfo->IsFitToFamilyMask(UI64LIT(0x0001800000800000), 0x00000040))
-                {
-                    m_canTrigger = true;
-                }
-                break;
-            case SPELLFAMILY_ROGUE:
-                // For poisons need do it
-                if (m_spellInfo->IsFitToFamilyMask(UI64LIT(0x000000101001E000)))
-                {
-                    m_canTrigger = true;
-                }
-                break;
-            case SPELLFAMILY_HUNTER:
-                // Hunter Rapid Killing/Explosive Trap Effect/Immolation Trap Effect/Frost Trap Aura/Snake Trap Effect/Explosive Shot
-                if (m_spellInfo->IsFitToFamilyMask(UI64LIT(0x0100200000000214), 0x200))
-                {
-                    m_canTrigger = true;
-                }
-                break;
-            case SPELLFAMILY_PALADIN:
-                // For Judgements (all) / Holy Shock triggers need do it
-                if (m_spellInfo->IsFitToFamilyMask(UI64LIT(0x0001000900B80400)))
-                {
-                    m_canTrigger = true;
-                }
-                break;
-            default:
-                break;
-        }
-    }
-
-    // Get data for type of attack and fill base info for trigger
-    switch (m_spellInfo->GetDmgClass())
-    {
-        case SPELL_DAMAGE_CLASS_MELEE:
-            m_procAttacker = PROC_FLAG_SUCCESSFUL_MELEE_SPELL_HIT;
-            if (m_attackType == OFF_ATTACK)
-            {
-                m_procAttacker |= PROC_FLAG_SUCCESSFUL_OFFHAND_HIT;
-            }
-            m_procVictim   = PROC_FLAG_TAKEN_MELEE_SPELL_HIT;
-            break;
-        case SPELL_DAMAGE_CLASS_RANGED:
-            // Auto attack
-            if (m_spellInfo->HasAttribute(SPELL_ATTR_EX2_AUTOREPEAT_FLAG))
-            {
-                m_procAttacker = PROC_FLAG_SUCCESSFUL_RANGED_HIT;
-                m_procVictim   = PROC_FLAG_TAKEN_RANGED_HIT;
-            }
-            else // Ranged spell attack
-            {
-                m_procAttacker = PROC_FLAG_SUCCESSFUL_RANGED_SPELL_HIT;
-                m_procVictim   = PROC_FLAG_TAKEN_RANGED_SPELL_HIT;
-            }
-            break;
-        default:
-            if (IsPositiveSpell(m_spellInfo->Id))           // Check for positive spell
-            {
-                m_procAttacker = PROC_FLAG_SUCCESSFUL_POSITIVE_SPELL;
-                m_procVictim   = PROC_FLAG_TAKEN_POSITIVE_SPELL;
-            }
-            else if (m_spellInfo->HasAttribute(SPELL_ATTR_EX2_AUTOREPEAT_FLAG))   // Wands auto attack
-            {
-                m_procAttacker = PROC_FLAG_SUCCESSFUL_RANGED_HIT;
-                m_procVictim   = PROC_FLAG_TAKEN_RANGED_HIT;
-            }
-            else                                           // Negative spell
-            {
-                m_procAttacker = PROC_FLAG_SUCCESSFUL_NEGATIVE_SPELL_HIT;
-                m_procVictim   = PROC_FLAG_TAKEN_NEGATIVE_SPELL_HIT;
-            }
-            break;
-    }
-
-    // some negative spells have positive effects to another or same targets
-    // avoid triggering negative hit for only positive targets
-    m_negativeEffectMask = 0x0;
-    for (int i = 0; i < MAX_EFFECT_INDEX; ++i)
-        if (!IsPositiveEffect(m_spellInfo, SpellEffectIndex(i)))
-        {
-            m_negativeEffectMask |= (1 << i);
-        }
-
-    // Hunter traps spells (for Entrapment trigger)
-    // Gives your Immolation Trap, Frost Trap, Explosive Trap, and Snake Trap ....
-    if (m_spellInfo->IsFitToFamily(SPELLFAMILY_HUNTER, UI64LIT(0x000020000000001C)))
-    {
-        m_procAttacker |= PROC_FLAG_ON_TRAP_ACTIVATION;
-    }
-}
-
-/**
- * @brief Clears all accumulated target lists and delay tracking.
- */
-void Spell::CleanupTargetList()
-{
-    m_UniqueTargetInfo.clear();
-    m_UniqueGOTargetInfo.clear();
-    m_UniqueItemInfo.clear();
-    m_delayMoment = 0;
-}
-
-/**
- * @brief Adds a unit target entry for a spell effect.
- *
- * @param pVictim The unit target.
- * @param effIndex The effect index being applied.
- */
-void Spell::AddUnitTarget(Unit* pVictim, SpellEffectIndex effIndex)
-{
-    SpellEffectEntry const *spellEffect = m_spellInfo->GetSpellEffect(effIndex);
-    if (!spellEffect || spellEffect->Effect == 0)
-    {
-        return;
-    }
-
-    // Check for effect immune skip if immuned
-    bool immuned = pVictim->IsImmuneToSpellEffect(m_spellInfo, effIndex, pVictim == m_caster);
-
-    if (pVictim->GetTypeId() == TYPEID_UNIT && ((Creature*)pVictim)->IsTotem() && (m_spellFlags & SPELL_FLAG_REDIRECTED))
-    {
-        immuned = false;
-    }
-
-    ObjectGuid targetGUID = pVictim->GetObjectGuid();
-
-    // Lookup target in already in list
-    for (TargetList::iterator ihit = m_UniqueTargetInfo.begin(); ihit != m_UniqueTargetInfo.end(); ++ihit)
-    {
-        if (targetGUID == ihit->targetGUID)                 // Found in list
-        {
-            if (!immuned)
-            {
-                ihit->effectMask |= 1 << effIndex;          // Add only effect mask if not immuned
-            }
-            return;
-        }
-    }
-
-    // This is new target calculate data for him
-
-    // Get spell hit result on target
-    TargetInfo target;
-    target.targetGUID = targetGUID;                         // Store target GUID
-    target.effectMask = immuned ? 0 : (1 << effIndex);      // Store index of effect if not immuned
-    target.processed  = false;                              // Effects not applied on target
-
-    // Calculate hit result
-    target.missCondition = m_caster->SpellHitResult(pVictim, m_spellInfo, m_canReflect);
-
-    // spell fly from visual cast object
-    WorldObject* affectiveObject = GetAffectiveCasterObject();
-
-    // Spell have speed (possible inherited from triggering spell) - need calculate incoming time
-    float speed = m_spellInfo->speed == 0.0f && m_triggeredBySpellInfo ? m_triggeredBySpellInfo->speed : m_spellInfo->speed;
-    if (speed > 0.0f && affectiveObject && (pVictim != affectiveObject || (m_targets.m_targetMask & (TARGET_FLAG_SOURCE_LOCATION | TARGET_FLAG_DEST_LOCATION))))
-    {
-        // calculate spell incoming interval
-        float dist = 0.0f;                                  // distance to impact
-        if (pVictim == affectiveObject)                     // Calculate dist to destination target also for self-cast spells
-        {
-            if (m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION)
-            {
-                dist = affectiveObject->GetDistance(m_targets.m_destX, m_targets.m_destY, m_targets.m_destZ);
-            }
-            else                                            // Must have Source Target
-            {
-                dist = affectiveObject->GetDistance(m_targets.m_srcX, m_targets.m_srcY, m_targets.m_srcZ);
-            }
-        }
-        else                                                // normal unit target, take distance
-        {
-            dist = affectiveObject->GetDistance(pVictim->GetPositionX(), pVictim->GetPositionY(), pVictim->GetPositionZ());
-        }
-
-        if (dist < 5.0f)
-        {
-            dist = 5.0f;
-        }
-        target.timeDelay = (uint64) floor(dist / speed * 1000.0f);
-
-        // Calculate minimum incoming time
-        if (m_delayMoment == 0 || m_delayMoment > target.timeDelay)
-        {
-            m_delayMoment = target.timeDelay;
-        }
-    }
-    // Spell casted on self - mostly TRIGGER_MISSILE code
-    else if (m_spellInfo->speed > 0.0f && affectiveObject && pVictim == affectiveObject)
-    {
-        float dist = 0.0f;
-        if (m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION)
-        {
-            dist = affectiveObject->GetDistance(m_targets.m_destX, m_targets.m_destY, m_targets.m_destZ);
-        }
-
-        target.timeDelay = (uint64) floor(dist / m_spellInfo->speed * 1000.0f);
-    }
-    else
-    {
-        target.timeDelay = UI64LIT(0);
-    }
-
-    // If target reflect spell back to caster
-    if (target.missCondition == SPELL_MISS_REFLECT)
-    {
-        // Calculate reflected spell result on caster
-        target.reflectResult =  m_caster->SpellHitResult(m_caster, m_spellInfo, m_canReflect);
-
-        if (target.reflectResult == SPELL_MISS_REFLECT)     // Impossible reflect again, so simply deflect spell
-        {
-            target.reflectResult = SPELL_MISS_PARRY;
-        }
-
-        // Increase time interval for reflected spells by 1.5
-        target.timeDelay += target.timeDelay >> 1;
-
-        m_spellFlags |= SPELL_FLAG_REFLECTED;
-    }
-    else
-    {
-        target.reflectResult = SPELL_MISS_NONE;
-    }
-
-    // Add target to list
-    m_UniqueTargetInfo.push_back(target);
-}
-
-/**
- * @brief Resolves and adds a unit target by guid for a spell effect.
- *
- * @param unitGuid The unit guid to resolve.
- * @param effIndex The effect index being applied.
- */
-void Spell::AddUnitTarget(ObjectGuid unitGuid, SpellEffectIndex effIndex)
-{
-    if (Unit* unit = m_caster->GetObjectGuid() == unitGuid ? m_caster : sObjectAccessor.GetUnit(*m_caster, unitGuid))
-    {
-        AddUnitTarget(unit, effIndex);
-    }
-}
-
-/**
- * @brief Adds a game object target entry for a spell effect.
- *
- * @param pVictim The game object target.
- * @param effIndex The effect index being applied.
- */
-void Spell::AddGOTarget(GameObject* pVictim, SpellEffectIndex effIndex)
-{
-    SpellEffectEntry const* spellEffect = m_spellInfo->GetSpellEffect(effIndex);
-    if (!spellEffect || spellEffect->Effect == 0)
-    {
-        return;
-    }
-
-    ObjectGuid targetGUID = pVictim->GetObjectGuid();
-
-    // Lookup target in already in list
-    for (GOTargetList::iterator ihit = m_UniqueGOTargetInfo.begin(); ihit != m_UniqueGOTargetInfo.end(); ++ihit)
-    {
-        if (targetGUID == ihit->targetGUID)                 // Found in list
-        {
-            ihit->effectMask |= (1 << effIndex);            // Add only effect mask
-            return;
-        }
-    }
-
-    // This is new target calculate data for him
-
-    GOTargetInfo target;
-    target.targetGUID = targetGUID;
-    target.effectMask = (1 << effIndex);
-    target.processed  = false;                              // Effects not apply on target
-
-    // spell fly from visual cast object
-    WorldObject* affectiveObject = GetAffectiveCasterObject();
-
-    // Spell can have speed - need calculate incoming time
-    float speed = m_spellInfo->speed == 0.0f && m_triggeredBySpellInfo ? m_triggeredBySpellInfo->speed : m_spellInfo->speed;
-    if (speed > 0.0f && affectiveObject && pVictim != affectiveObject)
-    {
-        // calculate spell incoming interval
-        float dist = affectiveObject->GetDistance(pVictim->GetPositionX(), pVictim->GetPositionY(), pVictim->GetPositionZ());
-        if (dist < 5.0f)
-        {
-            dist = 5.0f;
-        }
-        target.timeDelay = (uint64) floor(dist / speed * 1000.0f);
-        if (m_delayMoment == 0 || m_delayMoment > target.timeDelay)
-        {
-            m_delayMoment = target.timeDelay;
-        }
-    }
-    else
-    {
-        target.timeDelay = UI64LIT(0);
-    }
-
-    // Add target to list
-    m_UniqueGOTargetInfo.push_back(target);
-}
-
-/**
- * @brief Resolves and adds a game object target by guid for a spell effect.
- *
- * @param goGuid The game object guid to resolve.
- * @param effIndex The effect index being applied.
- */
-void Spell::AddGOTarget(ObjectGuid goGuid, SpellEffectIndex effIndex)
-{
-    if (GameObject* go = m_caster->GetMap()->GetGameObject(goGuid))
-    {
-        AddGOTarget(go, effIndex);
-    }
-}
-
-/**
- * @brief Adds an item target entry for a spell effect.
- *
- * @param pitem The item target.
- * @param effIndex The effect index being applied.
- */
-void Spell::AddItemTarget(Item* pitem, SpellEffectIndex effIndex)
-{
-    SpellEffectEntry const* spellEffect = m_spellInfo->GetSpellEffect(effIndex);
-    if (!spellEffect || spellEffect->Effect == 0)
-    {
-        return;
-    }
-
-    // Lookup target in already in list
-    for (ItemTargetList::iterator ihit = m_UniqueItemInfo.begin(); ihit != m_UniqueItemInfo.end(); ++ihit)
-    {
-        if (pitem == ihit->item)                            // Found in list
-        {
-            ihit->effectMask |= (1 << effIndex);            // Add only effect mask
-            return;
-        }
-    }
-
-    // This is new target add data
-
-    ItemTargetInfo target;
-    target.item       = pitem;
-    target.effectMask = (1 << effIndex);
-    m_UniqueItemInfo.push_back(target);
-}
-
-/**
  * @brief Checks whether required alive targets are present in the current target list.
  *
  * @return True if all required effects have a valid alive target; otherwise, false.
